# ActiveRecord::Relation#insert_all: support column mode for highest performance