@kirs 2024-07-22
Opened an alternative implementation: https://github.com/rails/rails/pull/52395