@fatkodima 2025-07-05
@kirs Do you think it is worth finishing this PR to get some performance improvements and usability improvements (no need to construct hashes)?